requirements.txt
arxiv==1.4.7
Flask==2.3.2
langchain==0.0.188
numpy==1.24.3
openai==0.27.5
pypdf==3.8.1
python-dotenv==1.0.0
